pid control basic throttle

+++ cmd/test-motor/motor.go
@@ -14,14 +14,14 @@ func main() {
 	motor := flag.Int("motor", 0, "motor")
 	flag.Parse()
 
-	const maxThrottle float32 = 10
+	const maxThrottle float64 = 10
 	const steps int = 10
-	var dThrottle float32 = maxThrottle / float32(steps)
-	var throttle float32 = 0
+	var dThrottle float64 = maxThrottle / float64(steps)
+	var throttle float64 = 0
 	esc := motors.NewESC()
 	esc.On()
 	time.Sleep(3 * time.Second)
-	throttles := map[uint8]float32{0: 0, 1: 0, 2: 0, 3: 0}
+	throttles := map[uint8]float64{0: 0, 1: 0, 2: 0, 3: 0}
 	for repeat := 0; repeat < 2; repeat++ {
 		for step := 0; step < steps; step++ {
 			fmt.Println("motor: ", *motor, ", throttle:  ", throttle, "%")
